# a7_openapi.yaml
openapi: 3.0.0
info:
  version: '1.0'
  title: 'LBAW Collaborative News Web API'
  description: 'Web Resources Specification (A7) for Collaborative News'
servers:
- url: http://lbaw2103.lbaw.fe.up.pt/
  description: Production server
...
paths:
...
 /admin/users:
    get:
      operationId: R403
      summary: 'R403: View Users Management Page'
      description: "View Users Management Page. Access: ADM"
      tags:
        - 'M04: User Administration and Static pages'

      responses:
        '200':
          description: 'Ok. Show UI17'
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    username:
                      type: string
                    email:
                      type: string
                    reputation:
                      type: integer
        '403':
          description: 'Forbidden'
...